\name{ml100k}
\alias{iranChurn}

\title{
MovieLens Data, 100K 
}

\description{

From
\url{https://www.kaggle.com/datasets/prajitdatta/movielens-100k-dataset}
}

\details{

\code{data}: Main data, in userID, itemID, rating and timestamp
format.

\code{user}: User demographics; \code{occupation} adds to it.

\code{item}: Movie info; see Kaggle site for details; \code{genre} add
to it.

}

